Fix Sonar Issues - [csharpsquid:S1450]
